ifc_vfs_dokan: fix open_directory callback
In case of the open_directory Dokan callback the target inode must be a
directory, otherwise the call must fail. Note that the opposite
(create_file just for files) is not true but Dokan unfortunately doesn't
pass the the FILE_NON_DIRECTORY_FILE flag and strips trailing
backslashes. This leaves us clueless and we need to hope that we don't
run into the corner cases.