vfs: invalidate folder cache when renaming entries
Renaming an entry in a folder doesn't change the folder itself but the
child document where the entry points to. As the titles of entries are
lazily updated (cached for one second) we need to invalidate the child
entry cache after renames. Otherwise an immediately following readdir
would return outdated filenames.