replicator_worker: use a set instead of a queue for backlog
The replicator is always replicating one document at a time with its
complete history. While traversing the history any referenced documents
are added to the backlog. With deep histories the same documents are
added many times to the backlog.

Previously a queue was used for the backlog which led to the situation
that most documents were handled over and over again on deep histories.
By using a set for the backlog each referenced document will only be
replicated once.