Make all Ids variable length
Right now all Ids (DId, RId, PId) are fixed to 128 bit and the whole
value range is used. Some Ids might need to be extended in the future
which would not be possible with this scheme.

By making the length of Ids variable it is easy to define new formats in
the future. Right now Ids stay at 128 bit whereas RIds and PIds are
extended to the full 192 bit SHA1 range.