Store comments as a dedicated field in revision
The comment about a revision is very special compared to other meta
data because it directly relates to the revision as entity, rather than to the
contents of the document. Also, the comment should be given explicitly
for every new revision and otherwise be reset.

Now the comment can be given when committing or suspending a document.
The comment will start empty for each newly created revision. The stat
operation on a revision can be used to read the committed comment
afterwards.