server: add crypt_store
The crypt_store is just a front end to any other store which encrypts
all data. The underlying store is mounted with the 'noverify' options
and won't see any meaningful data.

The actual encryption is done with AES where the master key is stored in
a LUKS like scheme. Given the right password the master key is decrypted
and checked against a hash sum to see if the password was correct. There
can be more than one password but adding a new one needs at leas one old
password to derive the master key.

Despite the actual data all the DIds, RIds, and PIds are encrypted too
to conceal if some document/revision/part is actually in the store. Also
all fields in the revisions are encrypted. Note that history graph of
documents/revisions is still visible which might constitute a side
channel for testing the existence of known documents.