Get rid of SysInfo client call
The SysInfo call was only used to get the base path of the FUSE/Dokan
mount path. As this path is not necessarily unique (in the future) and
because this is the only usage for SysInfo we get rid of it altogether
and replace it with a dedicated 'GetPath' call.