server: move test over from the client
The tiny test suite in the client is outdated and also on the wrong
level. Most of the tests there cover core PeerDrive functions which
should be tested in the server first.

The tests will be done by writing common_test test suites. They can be
extended gradually by re-implementing the test from the client unit
tests or when bugs are found.