file_store: only use absolute paths internally
The path to the store location in the file system might be given as a
relative path. But if the current working directory of the Erlag VM is
later changed to another directory the store will be writing to a false
location. Hence, always convert the path to an absolute name.