sync: improve resistance against store rollbacks
All added or changed documents in a store get a monotonically increasing
sequence number. This allows the synchronization from one store to the
other select any documents that have been changed (or added) since the
last sync point and ignore older ones.

The last sequence number was stored solely in the source store. This
approach has a number of problems:

 * The source store might be read-only for the user. This makes storing
   the sync anchor in the source store inconvenient. The destination
   store must be writable anyways.
 * If the destination store is restored to an older state (backup image,
   power loss) the synchronization will not be able to re-synchronize
   the potentially affected documents.

So instead of storing the sync anchor in the source store we now store
the anchor in both stores (if possible). The destination store is
authoritative, whereas the source store might provide a upper bound of
the anchor. Effectively the minimum sync anchor of either store is used.
This way any store could be rewound without disturbing the
synchronization. It is no error if the anchor cannot be updated in the
source store.