server: unify usage of handles
Any handle that is given back by a store is treated the same way for all
operations. First the handle is writable, then the changes are
committed, leaving the handle read-only, and finally the handle is
closed. Any resources that are prevented from garbage collection by the
handle are kept as long as the handle is open.

At the same time the peerdrive_store:put_rev_part/3 behaviour was
changed to always return a handle and that the revision itself is also
exempted from garbage collection. The rationale is that the caller wants
to have a certain revision in the store, thus its garbage collection
must be prevented even if its already there.

Finally, bump the netstore protocol major version as these changes are
backwards incompatible.