replicator: fix replication
Instead of replicating all requested elements from top to bottom we do
it in reverse, that is from bottom to top. Otherwise, if the replication
is canceled and re-done, not all elements would be replicated. Without
the 'verbose' flag the replication will not traverse any deeper as soon
as an element is already present on the destination. By doing the
replication from bottom up it will not stop prematurely anymore on the
2nd try.