Fix replication, 2nd part
Applications must replicate items first before creating a link to them.
Otherwise the container with the new link might get synchronized to
other stores but the actual target will not follow because it is just
being replicated and are not completely there yet.

Fix this behavior for the folder view/browser. It also required a change
in the client API to return a handle from the replicate operation to
prevent garbage collection of the replicated items until they are
successfully linked.