server: provide a path query service (WalkPathReq)
The WalkPath request can be used to query the PeerDrive server for a set
of documents. In it's basic form it is used very much like regular file
paths. The currently supported format is:

  label:[foo/bar/baz]

where 'label' is the label of any mounted store and the stuff in
brackets is an optional path traversal. The WalkPath request will return
a list of docs or revs because names are not necessarily unique per
directory in PeerDrive.

Though for the future the WalkPath request is likely to be extended into
something XPath like...