Start PeerDrive daemon per-user by default
The PeerDrive daemon is now started by an autostart .desktop file for
each user on login. It is still possible to start PeerDrive via init as
system daemon by adjusting /etc/default/peerdrive.

Unfortunately the daemon will still live after the logout of the user
because the integration with XSMP/ConsoleKit is lacking. In the long run
systemd user sessions should handle it, just like launchd does already
on OSX.