Ephemeral revisions
Ephemeral revisions are just like any other revision except that their
parent links are weak, that is they do not prevent garbage collecion of
their ancestors. As such, the history of these revisions will be trunced
with every garbage collector run. Such revisions are usefull to track
files in PeerDrive that are not important such as:

 * Caches (thumbs.db)
 * Special system files (desktop.ini, .ds_store)
 * Temporary files (foo.txt~)

As a side effect ephemeral revisions cannot be merged effectively
because they loose their history quite fast. In fact the merger and
replicator stop at ephemeral revisions because it makes no sense to
merge and/or replicate them.