Extend object model with more timestamps
Additionally to the mtime field in the revision the following time
fields are added:

 - crtime: creation time of the document
 - attachment crtime: creation time of the attachment
 - attachment mtime: modification time of the attachment

And while we are at it remove the stupid distinction between #revision
and #rev_stat records. Both records contain almost the same information
(except the data/attachment sizes). Get rid of that and introduce a
common #rev record.